Enable show route for a JSON request for a specific Link
We enable the show route for nested Interactions, so the link for the specific ServiceInteraction can be requested.
The request would look like:

`GET /local_authorities/bolton/services/412/interactions/8.json`

The most contentious part of this implementation is probably that we have to convert from LGSL and LGIL codes to slugs in the controller. Frontend will make the request and only knows about LGSL and LGIL codes. However the routes we have set up in LocalLinksManager currently all work with slugs. That's why we need to convert the params hash in the controller to convert the numbers into relevant slugs.