Add initial dashboard homepages for V2
Pretty scrappy, but they do the job.  The "all broken links" page is
fairly slow, but without working out what to do for pagination it's
not clear how to speed it up.

We also don't have the real broken link counts on the pages.  We'll
try to get those next.