Add a dropdown for choosing sort order
Options for this all come from a hash keyed on the param value.  To
avoid nils we set the value of the 'default' key to be what is returned
by the hash by default when an unknown key is requested.  Bit overkill
for a prototype really.  We can use the hash to build the dropdown
in the view though, so it's somewhat useful.

We'll add the sorting by broken link counts into the next commit as
this will require more complex queries.