Add support for sorting by broken link count
The queries here are fairly complex, and we've had to drop down to use
Arel directly, rather than the more limited (but clearer) interface that
activerecord provides for building queries.

Ultimately we select all the LocalAuthorities or Services and add a count
of broken links by joining on a subquery, which is all the broken links
grouped by the local authority or service they belong to.  The service
version is made even more complex because we have to join through
service interactions to get this.

Happily though, the hash for sort options and order_args param is enough
to support this (albeit with a more complex starting point).

By selecting the count from the subquery into the activerecord query for
the local authority or service it gets added as an attribute to each
instance returned, meaning we can finally drop it onto the page and
make the numbers real.  Which is nice.