Add the 2nd filter control for filtering the current page
The pages all use the same UI: a select2-based dropdown lets you
search for a things you can filter the page by.  Once selected the
rows that do not match this thing are hidden.  If the dropdown is
cleared we show all the rows again.

The services and local authorities pages are straightforward, but
the broken links page allows for two different things to be searched
for and has two types of rows to toggle which makes things more complex.

There's probably scope for making this more componentized so it can
be configured and shared amongst all 3 pages, but not now.