Add link type filtering for v2 local authorities and services indexes
These filters work similarly to the sort_order param and run via a
request/response rather than filtering the data on the page.  This
does leave us in the situation where the sort and link type controls
work via http, but the dropdown filter control works on the page and
can't be refelcted in the URL for sharing purposes.  A more robust
version of this should be developed for production.

The filters are implemented as radio buttons wrapped in links, to trigger
the navigation when the user clicks the button rather than the text
we have to track the change event and navigate them manually.  If the
whole control panel was a form we could just submit normally.