Add v2 edit page
We copy most of the html and controller code from the "v1" version. There
are enough minor differences in the html and behaviour to warrant a brand
new version here though.  The key difference is in where the cancel button
and successful save will send you back to.  In here we record the referer
when we hit the edit url and send us back there.  Compare this with the
original version where there's only one route in so we always go back to
the same place.