Redirect to starting point from authority edit page
We use the same thing we added to the v2 links controller, but there's
no need to have a whole new v2 controller as this is the only difference
and it's useful in the v1 context too.

We also add a shared flashes partial so we can see the success messages
that are set when homepage or other links are edited.