Allow searching for a link via dropdown
We rely on select2 tagging and tag creation to do this.  If the term
being typed doesn't match select2 invokes the `createTag` function. We
look at the value, and if it starts with `http://` or `https://` then
we filter out rows where the url doesn't match what is typed.

There's some issues with this where we hide the first row for an
interaction, but we can tidy it up.  It's also much faster to show/hide
the rows because we're using css, not iterating manually.