Update total counts and row counts as we filter on the page
When the filters on the page update, make sure that the on screen
numbers also update.  For those indexes (broken links + services)
that display intermediate counts, also update those rows too.

We also change the filtering JS to not treat the la header rows
on the broken links + services as special cases, only data rows will
be hidden or shown, and we leave it up to the pages to handle their
parent rows manually.  This means they no longer need to treat the
'local-authority-id' selectedType as special case.