Mostly styling tweaks in response to design snagging
We also extract the whole panel heading as a shared partial because it
was getting pretty complex with the extra styling tweaks needed to reduce
the margins and padding on the big numbers.  It's a bit odd in that we
rely on either a simple form for the numbers, passed in as arguments, or
the calling template setting a content_for.  This isn't super obvious and
should be avoided if we go to production with this.

Adding the padding to the tables seems to knock them out of the panel on
the broken links page. I suspect this is because that panel is constrained
to a 9 grid unit width, whereas the others are the full width 12 grid unit
width.