Convert sidebar nav pills and convert to a tab nav
The rationale for this is that the table really wants to have as much
space as it can get to show urls, but sacrificing 3 grid units to the
sidebar nav means the table can't really breathe.  If we make it a top
nav instead the table on these pages gets to be full-width and have more
space.  This has the advantage of making all our tables full-width so
we get some consistency.

The choice of nav tabs was made for simplicity reasons, it may not be
the final design choice.

We also extract the nav to a shared partial to make it easier to change
across all 3 views.