Show full URLs in link tables and fix layout
By telling the tables to have fixed layout the size of each column is
based on the first row. We use bootstrap grid classes to set the widths
giving the link cell the largest share. We also tell the link cell that
it is allowed to break text inside words too, and this allows us to write
the entire url to the page as it will wrap, rather than make the table
really wide, or overflow into other cells.