Keep filter-dropdown value on pressing back
When the user selects something in the filter drop-down and then nagivates
away and comes back, the filter-dropdown will retain the selected item so
we need to filter the on page results if we can.  We get a 'pageshow'
event in this circumstance (on every page load, be it the initial render
or a re-render from back) so we can run the item callback manually by
constructing enough of the event data to be able to do so.