Adds search tracking for index pages
This will fire a `filterUsed` event when a user writes something in the filter search box
(action: `filterUsed`, category: `userInteraction:LLM`)

This commit also sets a custom dimension to be sent along with the GA events being sent in
LocalLinksManager (custom dimension index: 7, name: `Scorecard Total`). The Scorecard total
is a value that indicates the number of links being displayed on an index page. It is sent
when the index pages first loads, using the `auto_track_event` module.